+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Bit count to limit".
 - Tests for "Porcupine fever".
 - Tests for "Bijective numeration".
+- Tests for "Plight of the fellowship of the ring".
 
 ### Fixed
 - Tests groups for "CGFunge interpreter".